Rename sketch to match folder name
The Arduino IDE requires that a sketch be located in a folder of the same name. On a filename case insensitive OS like Windows, ArduBot and arduBot are considered identical. However, on a filename case sensitive operating system such as Linux they are not. When opening a file that does not meet this requirement the Arduino IDE presents a dialog:

The file "arduBot.ino" needs to be inside a sketch folder named "arduBot".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file arduBot.ino to the new folder, leaving behind pitches.h. This causes compilation of the sketch to fail:

arduBot\arduBot.ino:28:21: fatal error: pitches.h: No such file or directory

 #include "pitches.h"